# core/core-api/src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: soo-auth-service API
  description: 인증 서비스 API 가이드 (REST API Docs).
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost
...
paths:
...
  /api/v1/auth/local/phone-verifications/confirm:
    post:
      tags:
      - Local Phone
      summary: 휴대폰 인증번호 확인
      description: "인증번호를 검증하고 회원가입에 사용할 phoneVerificationToken을 발급합니다.\n\n### 인증\n\
        - 인증 불필요\n\n### 요청 바디 필드\n- `phoneNumber`: 인증 확인할 휴대폰 번호\n- `verificationId`:\
        \ 인증 요청 식별자\n- `code`: 6자리 인증번호\n\n### 응답 바디 필드\n- `result`: 요청 처리 결과 (SUCCESS/ERROR)\n\
        - `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`: 요청 추적 ID (있을\
        \ 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`: 요청 경로\n-\
        \ `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열 (없으면 null)\n\
        - `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로케일 (있을 때만 반환)\n\
        - `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 인증 확인 결과\n- `data.phoneVerificationToken`:\
        \ 회원가입에 사용할 휴대폰 인증 토큰\n- `data.expiresInSec`: 인증 토큰 만료까지 남은 시간(초)\n\n### 공\
        통 에러\n- **401** 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation\
        \ 실패(필드 형식/필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과\
        로 자동 포함됩니다."
      operationId: auth-local-phone-verification-confirm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-local-phone-verifications-confirm168401481'
            examples:
              auth-local-phone-verification-confirm:
                value: "{\n  \"phoneNumber\" : \"+82 10-1234-5678\",\n  \"verificationId\"\
                  \ : \"verification-id-001\",\n  \"code\" : \"123456\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-local-phone-verifications-confirm445202834'
              examples:
                auth-local-phone-verification-confirm:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:52.0053811\",\n    \"request\" : {\n  \
                    \    \"path\" : \"/api/v1/auth/local/phone-verifications/confirm\"\
                    ,\n      \"method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n\
                    \    \"phoneVerificationToken\" : \"verified-phone-token\",\n\
                    \    \"expiresInSec\" : 600\n  }\n}"
  /api/v1/auth/local/phone-verifications/request:
    post:
      tags:
      - Local Phone
      summary: 휴대폰 인증번호 발급
      description: "휴대폰 번호로 인증번호를 발급하고 verificationId를 반환합니다.\n\n### 인증\n- 인증 불필요\n\
        \n### 요청 바디 필드\n- `phoneNumber`: 인증 요청할 휴대폰 번호\n\n### 응답 바디 필드\n- `result`:\
        \ 요청 처리 결과 (SUCCESS/ERROR)\n- `meta.timestamp`: 응답 생성 시각 (서버 로컬 시간)\n- `meta.requestId`:\
        \ 요청 추적 ID (있을 때만 반환)\n- `meta.durationMs`: 처리 시간(ms)\n- `meta.request.path`:\
        \ 요청 경로\n- `meta.request.method`: HTTP 메서드\n- `meta.request.query`: 쿼리 문자열\
        \ (없으면 null)\n- `meta.paging`: 페이징 메타 정보 (있을 때만 반환)\n- `meta.locale`: 응답 로\
        케일 (있을 때만 반환)\n- `error`: 오류 정보 (에러 응답에서만 포함)\n- `data`: 인증 요청 결과\n- `data.verificationId`:\
        \ 인증 요청 식별자\n- `data.expiresInSec`: 인증번호 만료까지 남은 시간(초)\n\n### 공통 에러\n- **401**\
        \ 인증 실패/만료/위조\n- **403** 권한 없음(권한 필요한 API)\n- **400** validation 실패(필드 형식\
        /필수 누락)\n\n> 실제 요청/응답 예시는 Swagger UI의 **Examples** 섹션에 테스트 실행 결과로 자동 포함됩니다\
        ."
      operationId: auth-local-phone-verification-request
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-auth-local-phone-verifications-request-394943525'
            examples:
              auth-local-phone-verification-request:
                value: "{\n  \"phoneNumber\" : \"+82 10-1234-5678\"\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-auth-local-phone-verifications-request271430078'
              examples:
                auth-local-phone-verification-request:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"meta\" : {\n    \"timestamp\"\
                    \ : \"2026-04-06T10:27:51.959482695\",\n    \"request\" : {\n\
                    \      \"path\" : \"/api/v1/auth/local/phone-verifications/request\"\
                    ,\n      \"method\" : \"POST\"\n    }\n  },\n  \"data\" : {\n\
                    \    \"verificationId\" : \"verification-id-001\",\n    \"expiresInSec\"\
                    \ : 180\n  }\n}"
...
components:
  schemas:
    api-v1-auth-local-phone-verifications-confirm168401481:
      required:
      - code
      - phoneNumber
      - verificationId
      type: object
      properties:
        code:
          type: string
          description: 6자리 인증번호
        phoneNumber:
          type: string
          description: 인증 확인할 휴대폰 번호
        verificationId:
          type: string
          description: 인증 요청 식별자
...
    api-v1-auth-local-phone-verifications-request271430078:
      required:
      - data
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 처리 결과 (SUCCESS/ERROR)
        data:
          required:
          - expiresInSec
          - verificationId
          type: object
          properties:
            expiresInSec:
              type: number
              description: 인증번호 만료까지 남은 시간(초)
            verificationId:
              type: string
              description: 인증 요청 식별자
          description: 인증 요청 결과
        meta:
          required:
          - timestamp
          type: object
          properties:
            request:
              required:
              - method
              - path
              type: object
              properties:
                path:
                  type: string
                  description: 요청 경로
                method:
                  type: string
                  description: HTTP 메서드
                query:
                  type: string
                  description: 쿼리 문자열 (없으면 null)
                  nullable: true
            requestId:
              type: string
              description: 요청 추적 ID (있을 때만 반환)
              nullable: true
            paging:
              type: object
              description: 페이징 메타 정보 (있을 때만 반환)
              nullable: true
            locale:
              type: string
              description: 응답 로케일 (있을 때만 반환)
              nullable: true
            durationMs:
              type: number
              description: 처리 시간(ms)
              nullable: true
            timestamp:
              type: string
              description: 응답 생성 시각 (서버 로컬 시간)
        error:
          type: object
          description: 오류 정보 (에러 응답에서만 포함)
          nullable: true
...
    api-v1-auth-local-phone-verifications-confirm445202834:
      required:
      - data
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 처리 결과 (SUCCESS/ERROR)
        data:
          required:
          - expiresInSec
          - phoneVerificationToken
          type: object
          properties:
            phoneVerificationToken:
              type: string
              description: 회원가입에 사용할 휴대폰 인증 토큰
            expiresInSec:
              type: number
              description: 인증 토큰 만료까지 남은 시간(초)
          description: 인증 확인 결과
        meta:
          required:
          - timestamp
          type: object
          properties:
            request:
              required:
              - method
              - path
              type: object
              properties:
                path:
                  type: string
                  description: 요청 경로
                method:
                  type: string
                  description: HTTP 메서드
                query:
                  type: string
                  description: 쿼리 문자열 (없으면 null)
                  nullable: true
            requestId:
              type: string
              description: 요청 추적 ID (있을 때만 반환)
              nullable: true
            paging:
              type: object
              description: 페이징 메타 정보 (있을 때만 반환)
              nullable: true
            locale:
              type: string
              description: 응답 로케일 (있을 때만 반환)
              nullable: true
            durationMs:
              type: number
              description: 처리 시간(ms)
              nullable: true
            timestamp:
              type: string
              description: 응답 생성 시각 (서버 로컬 시간)
        error:
          type: object
          description: 오류 정보 (에러 응답에서만 포함)
          nullable: true
...
    api-v1-auth-local-phone-verifications-request-394943525:
      required:
      - phoneNumber
      type: object
      properties:
        phoneNumber:
          type: string
          description: 인증 요청할 휴대폰 번호
...